+++ install.ps1
@@ -1,6 +1,6 @@
 #!/usr/bin/env pwsh
 
-$version = 'v0.11'
+$version = 'v0.12'
 $APP = 'maple'
 $url = "https://github.com/liuchengxu/vim-clap/releases/download/$version/$APP-"
 $output = "$PSScriptRoot\bin\$APP.exe"